Fix incorrect ProcessorStatus values
According to documentation, bit 5 is ignored and the overflow and
negative flags are actually bit 6 and 7 respectively.

Also, use bit shifts instead of plain integers to match the convention
used in various sources of documentation, and to make the mapping from
each enum value to a bit in the Processor Status register more obvious.