Implement opcodes used by the 3 * 10 test program
My design to separate fetching from the actual operations seems to fall
apart a bit because some operations want the address while others want
the value, even for the same fetch type.

My solution is to pass around a struct which specifies whether it is an
address or a value. The operation function can grab the data from memory
if its given an address but wants a value.

Not ideal to have some of the fetching happen in the operation itself,
but this was the cleanest way I could think of without resorting to
additional state in the CPU class, or worse, implementing each opcode as
its own function.

This could also be used for validation in the future, e.g. if a certain
operation requires an address but is given a value, or vice versa.