Create CPU and Memory separately to System
Rather than creating the CPU and Memory inside the System constructor,
create them separately and pass them in as arguments.

The main benefit here is to decouple the components so that they can be
tested individually, rather than having to go through a System.

This also allows usage of shared_ptr everywhere without requiring
enabled_shared_from_this then failing a call to shared_from_this()
in a constructor. Calling shared_from_this() in a constructor is invalid
because the enclosing shared_ptr is not yet fully constructed. Storing a
raw pointer to 'this' is perfectly perfectly valid, as long as members
are accessed outside of the constructor.