Allow creation of an empty ROM and Cartridge
This adds a ROM::Load() overload (i.e. with no parameters) which will
initialise two empty blocks of 16K each, one for PRG ROM and the other
for CHR ROM.

At the moment, writes to this memory are allowed, which is useful for
simple test cases.

This also adds the correct memory mapping to System so that reads and
writes are directed to the correct component. This still isn't ideal
since we're creating multiple, non-contiguous, blocks of memory.