Add SDL3 and Dear Imgui as dependencies
For now, these will be used for a visual display of the state of the CPU
and memory to help debugging.

I plan on using SDL3 for input, video and sound. However, I'm not too
keen on using Dear Imgui as the main UI so I'll pick something else when
the time comes.

Note that Dependencies.cmake has to specify everything related to SDL3
and Imgui in a specific order. Imgui doesn't have its own CMake file,
and so we have to create the target and then specify the header include
path for SDL3 to ensure that Imgui's SDL3 backend can build.