Create a debug UI with SDL3 and Dear Imgui
Use the SDL3 backend provided by Dear Imgui, and create a UI to load ROM
files as well as view the state of the CPU and memory.

The memory view goes via the system bus, so should accurately reflect
what the CPU and other components see.

SDL3 added SDL_ShowOpenFileDialog() - that's *incredibly* convenient!