Add debug UI to manually edit memory
This also makes it simple to directly initialise the ROM memory with a
small test program. Now I can step through the programs in the test
suite to narrow down why it's failing. I can't believe I wrote all of
the existing tests without this!!

Also, display the hex value of the opcode alongside the name, to make it
much easier to identify why the CPU is trying to execute an invalid one.