Increment PC immediately after fetching the opcode
All of the fetch functions now post increment the PC, fetch_relative()
can remove an off-by-one hack, and many of the branching instructions no
longer need to increment the PC in the non-branch cases.

This is exactly how you'd expect it to work, and the code is simpler.

I'm not 100% sure whether the branching instructions are correct, as
some of the test cases are failing now. But it's very likely that some
of the test cases have always been invalid.